cleanup: separate utils into separate files
