+++ tests/unit/index.test.ts
@@ -1,1 +1,2 @@
-import './util.test';
+import './utils/object.test';
+import './utils/path.test';
